pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = "Car data downloader and a backend service for the saved car data"
authors = ["Fasih Burak <fasihburak@gmail.com>"]
readme = "README.md"
license = "MIT"
classifiers = ["Framework :: Django", ]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.4"
djangorestframework = "^3.14.0"
requests = "^2.28.1"
beautifulsoup4 = "^4.11.1"
django-filter = "^22.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
